+++ src/main/java/com/yi/enhancement/mapper/ArticleTagRelationMapper.java
@@ -26,7 +26,7 @@ public interface ArticleTagRelationMapper extends BaseMapper<ArticleTagRelation>
      * @param tagId 标签Id
      * @return 所有文章Id
      */
-    List<Long> listArticleId(Long tagId);
+    List<Integer> listArticleId(Integer tagId);
 
     /**
      * 根据文章id获取该文章的所有标签
@@ -34,7 +34,7 @@ public interface ArticleTagRelationMapper extends BaseMapper<ArticleTagRelation>
      * @param articleId 文章id
      * @return tag列表
      */
-    List<Tag> listTagByArticleId(Long articleId);
+    List<Tag> listTagByArticleId(Integer articleId);
 
     /**
      * 根据文章Id查出所有的文章标签
@@ -42,7 +42,7 @@ public interface ArticleTagRelationMapper extends BaseMapper<ArticleTagRelation>
      * @param articleIds 文章id
      * @return 文章和标签的关系DTO
      */
-    List<ArticleTagRelationDTO> listArticleTagRelation(@Param("articleIds") List<Long> articleIds);
+    List<ArticleTagRelationDTO> listArticleTagRelation(@Param("articleIds") List<Integer> articleIds);
 
     /**
      * 根绝文章Id查出所有的文章标签
@@ -50,12 +50,12 @@ public interface ArticleTagRelationMapper extends BaseMapper<ArticleTagRelation>
      * @param articleId 文章id
      * @return 文章和标签的关系
      */
-    List<ArticleTagRelation> listArticleTagRelationByArticleId(@Param("articleId") Long articleId, @Param("deleted") Integer deleted);
+    List<ArticleTagRelation> listArticleTagRelationByArticleId(@Param("articleId") Integer articleId, @Param("deleted") Integer deleted);
 
     /**
      * 根据文章id删除文章标签
      *
      * @param articleId 文章id
      */
-    void deleteReally(@Param("articleId") Long articleId);
+    void deleteReally(@Param("articleId") Integer articleId);
 }
